% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve_model_cv.R
\name{add_interact_num}
\alias{add_interact_num}
\title{Add interaction numbers for panel data}
\usage{
add_interact_num(d)
}
\arguments{
\item{d}{data.frame of panel data}
}
\value{
Returns a vector specifying interactions
}
\description{
\code{add_interact_num} takes in data and returns a vector of interactions
}
