% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class_def.R
\name{check_ga_fsm}
\alias{check_ga_fsm}
\title{A function to check for errors}
\usage{
check_ga_fsm(object)
}
\arguments{
\item{object}{S4 ga_fsm object}
}
\description{
A function to check for errors
}

