% Generated by roxygen2 (4.0.2): do not edit by hand
\name{decode_action_vec}
\alias{decode_action_vec}
\title{Decodes Action Vector}
\usage{
decode_action_vec(string, states, inputs, actions)
}
\arguments{
\item{string}{Numeric vector.}

\item{states}{Numeric vector with the number of states.}

\item{inputs}{Numeric vector length one.}

\item{actions}{Numeric vector with the number of actions.}
}
\value{
Returns numeric vector.
}
\description{
\code{decode_action_vec} decodes action vector.
}
\details{
add here
}

