% Generated by roxygen2 (4.0.2): do not edit by hand
\name{decode_state_mat}
\alias{decode_state_mat}
\title{Decodes State Matrix}
\usage{
decode_state_mat(string, states, inputs, actions)
}
\arguments{
\item{string}{Numeric vector.}

\item{states}{Numeric vector with the number of states.}

\item{inputs}{Numeric vector length one.}

\item{actions}{Numeric vector with the number of actions.}
}
\value{
Returns numeric matrix.
}
\description{
\code{decode_state_mat} decodes action vector.
}
\details{
add here
}

