% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_def.R
\docType{methods}
\name{dotchart,ga_fsm-method}
\alias{dotchart,ga_fsm-method}
\title{Plots ga_fsm S4 object's variable importances}
\usage{
\S4method{dotchart}{ga_fsm}(x, labels)
}
\arguments{
\item{labels}{vector of labels for each point. For vectors the default is to
use names(x) and for matrices the row labels dimnames(x)[[1]].}
}
\description{
Plots ga_fsm S4 object's variable importances
}

