% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_def.R
\name{estimation_details}
\alias{estimation_details}
\title{Extracts slot relevant to estimating the fsm}
\usage{
estimation_details(x)
}
\arguments{
\item{x}{S4 ga_fsm object}
}
\description{
Extracts slot relevant to estimating the fsm
}

