% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fitnessC}
\alias{fitnessC}
\title{Fitness function for Prisoner Dilemma game data.}
\usage{
fitnessC(action_vec, state_mat, x)
}
\arguments{
\item{action_vec}{Integer Vector.}

\item{state_mat}{IntegerMatrix.}

\item{x}{IntegerMatrix.}
}
\description{
Fitness function for Prisoner Dilemma game data.
}

