% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fitnessCPP}
\alias{fitnessCPP}
\title{Fitness Function in C++}
\usage{
fitnessCPP(action_vec, state_mat, covariates, period)
}
\arguments{
\item{action_vec}{Integer Vector.}

\item{state_mat}{Integer Matrix.}

\item{covariates}{Integer Matrix.}

\item{period}{Integer Vector.}
}
\description{
Fitness Function in C++
}

