% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{fsm}
\alias{fsm}
\alias{fsm-package}
\title{fsm: A package for estimating FSM models.}
\description{
The fsm package provides one primary function for estimating a fsm decision
model and helper functions.
}
\section{FSM functions}{
 The fsm functions ...
}

