% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_def.R
\docType{methods}
\name{predict,ga_fsm-method}
\alias{predict,ga_fsm-method}
\title{Predicts new data with estimated model}
\usage{
\S4method{predict}{ga_fsm}(object, data, type = "prob",
  na.action = stats::na.omit, ...)
}
\arguments{
\item{data}{data.frame that has columns named "period" and "outcome" (period 
is the time period that the outcome action was taken), and the rest of the 
columns are predictors, ranging from one to three predictors. All of the 
(3-5 columns) should be named. The period and outcome columns should be 
integer vectors and the columns with the predictor variable data should be 
logical vectors (\code{TRUE, FALSE}). If the predictor variable data is not 
logical, it will coerced to logical with \code{base::as.logical()}.}

\item{type}{Not currently used.}

\item{na.action}{Optional function.}
}
\description{
Predicts new data with estimated model
}

