% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.summary.ga_fsm}
\alias{print.summary.ga_fsm}
\title{Prints a summary of a ga_fsm S4 object}
\usage{
print.summary.ga_fsm(x, digits = 3, ...)
}
\arguments{
\item{x}{ga_fsm object}

\item{digits}{numeric vector length one for how many significant digits to
print}

\item{...}{ignored.}
}
\description{
Prints a summary of a ga_fsm S4 object
}

