% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.ga_fsm}
\alias{summary.ga_fsm}
\title{Turns ga_fsm S4 object into list for printing}
\usage{
summary.ga_fsm(object, ...)
}
\arguments{
\item{object}{S4 ga_fsm object}

\item{...}{ignored}
}
\description{
Turns ga_fsm S4 object into list for printing
}

