% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/var_imp.R
\name{var_imp}
\alias{var_imp}
\title{Variable Importance Measure for A FSM Model}
\usage{
var_imp(state_mat, action_vec, data, outcome, period)
}
\arguments{
\item{state_mat}{Numeric matrix with rows as states and columns as
predictors.}

\item{action_vec}{Numeric vector indicating what action to take for each
state.}

\item{data}{Data frame that has "period" and "outcome" columns and rest of
cols are predictors, ranging from one to three predictors. All of the (3-5
columns) should be named.}

\item{outcome}{Numeric vector same length as the number of rows as data.}

\item{period}{Numeric vector same length as the number of rows as data.}
}
\value{
Numeric vector the same length as the number of columns of the
  provided state matrix (the number of predictors in the model) with relative
  importance scores for each predictor.
}
\description{
\code{var_imp} calculates the importance of the covariates of the model.
}
\details{
Takes the state matrix and action vector from an already evolved model and
the fitness function and data used to evolve the model (or this could be test
data), flips the values of each of the elements in the state matrix and
measures the change in fitness (prediction of data) relative to the original
model. Then these changes are summed across columns to provide the importance
of each of the columns of the state matrix.
}

