% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datafsm.R
\docType{package}
\name{datafsm}
\alias{datafsm}
\alias{datafsm-package}
\title{datafsm: A package for estimating FSM models.}
\description{
The datafsm package provides one primary function for estimating a fsm decision
model and helper functions.
}
\section{Data FSM functions}{
 The fsm functions ...
}

