% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafsm.R
\docType{package}
\name{datafsm}
\alias{datafsm}
\alias{datafsm-package}
\title{datafsm: A package for estimating FSM models.}
\description{
It relies on the \strong{GA} package:
Luca Scrucca (2013). GA: A Package for Genetic Algorithms in R.
Journal of Statistical Software, 53 (4), 1-37.
URL \url{https://www.jstatsoft.org/v53/i04/}.
}
\section{datafsm functions}{

\code{datafsm}'s main function for estimating a fsm decision
model:
\enumerate{
\item \code{\link{evolve_model}}
}
\code{datafsm}'s helper functions:
\enumerate{
\item \code{\link{evolve_model_cv}}
\item \code{\link{var_imp}}
\item \code{\link{decode_state_mat}}
\item \code{\link{decode_action_vec}}
\item \code{\link{fitnessCPP}}
\item \code{\link{build_bitstring}}
\item \code{\link{compare_fsm}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://jonathan-g.github.io/datafsm/}
  \item \url{https://github.com/jonathan-g/datafsm}
  \item Report bugs at \url{https://github.com/jonathan-g/datafsm/issues}
}

}
\author{
\strong{Maintainer}: Jonathan M. Gilligan \email{jonathan.gilligan@vanderbilt.edu} (\href{https://orcid.org/0000-0003-1375-6686}{ORCID})

Authors:
\itemize{
  \item John J. Nay \email{john.j.nay@gmail.com}
}

}
