% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ga_fsm-class}
\alias{ga_fsm-class}
\alias{print,ga_fsm-method}
\alias{show,ga_fsm-method}
\alias{summary,ga_fsm-method}
\title{An S4 class to return the results of using a GA to estimate a FSM with
evolve_model()}
\usage{
\S4method{print}{ga_fsm}(x, ...)

\S4method{show}{ga_fsm}(object)

\S4method{summary}{ga_fsm}(object, ...)
}
\arguments{
\item{object}{S4 ga_fsm object}
}
\description{
An S4 class to return the results of using a GA to estimate a FSM with
evolve_model()
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: An S4 method for printing a ga_fsm S4 object

\item \code{show}: An S4 method for showing a ga_fsm S4 object

\item \code{summary}: An S4 method for summarizing a ga_fsm S4 object
}}
\section{Slots}{

\describe{
\item{\code{call}}{Language from the call of the function evolve_model().}

\item{\code{actions}}{Numeric vector with the number of actions.}

\item{\code{states}}{Numeric vector with the number of states.}

\item{\code{GA}}{S4 object created by ga() from the GA package.}

\item{\code{state_mat}}{Numeric matrix with rows as states and columns as predictors.}

\item{\code{action_vec}}{Numeric vector indicating what action to take for each
state.}

\item{\code{predictive}}{Numeric vector of length one with test data accuracy if test
data was supplied; otherwise, a characte vector with a message that the
user should provide test data for better estimate of performance.}

\item{\code{degeneracy}}{List with message and sparse matrix.}

\item{\code{varImp}}{Numeric vector same length as number of columns of state matrix
with relative importance scores for each predictor.}
}}

