% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class_def.R
\docType{class}
\name{ga_fsm-class}
\alias{ga_fsm-class}
\alias{plot,ga_fsm-method}
\alias{print,ga_fsm-method}
\alias{show,ga_fsm-method}
\alias{summary,ga_fsm-method}
\title{An S4 class to return the results of using a GA to estimate a FSM with
\code{\link{evolve_model}}.}
\usage{
\S4method{print}{ga_fsm}(x, ...)

\S4method{show}{ga_fsm}(object)

\S4method{summary}{ga_fsm}(object, digits = 3)

\S4method{plot}{ga_fsm}(x, y, action_label)
}
\arguments{
\item{x}{S4 ga_fsm object}

\item{...}{ignored}

\item{object}{S4 ga_fsm object}

\item{digits}{Optional numeric vector length one for how many significant digits to
print, default is 3.}

\item{y}{not used.}

\item{action_label}{optional character vector same length as action vector, where each ith element corresponds to what that ith element in the action vector represents. This will be used to fill in the states (circles) of the state transition matrix to be plotted.}

\item{x}{S4 ga_fsm object}
}
\description{
An S4 class to return the results of using a GA to estimate a FSM with
\code{\link{evolve_model}}.

Turns ga_fsm S4 object into list of summaries for printing and then prints it.

Plots ga_fsm S4 object's state transition matrix
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: An S4 method for printing a ga_fsm S4 object

\item \code{show}: An S4 method for showing a ga_fsm S4 object

\item \code{summary}: An S4 method for summarizing a ga_fsm S4 object

\item \code{plot}: An S4 method for summarizing a ga_fsm S4 object
}}
\section{Slots}{

\describe{
\item{\code{call}}{Language from the call of the function \code{\link{evolve_model}}.}

\item{\code{actions}}{Numeric vector with the number of actions.}

\item{\code{states}}{Numeric vector with the number of states.}

\item{\code{GA}}{S4 object created by ga() from the GA package.}

\item{\code{state_mat}}{Numeric matrix with rows as states and columns as predictors.}

\item{\code{action_vec}}{Numeric vector indicating what action to take for each
state.}

\item{\code{predictive}}{Numeric vector of length one with test data accuracy if test
data was supplied; otherwise, a character vector with a message that the
user should provide test data for better estimate of performance.}

\item{\code{degeneracy}}{List with message and sparse matrix.}

\item{\code{varImp}}{Numeric vector same length as number of columns of state matrix
with relative importance scores for each predictor.}

\item{\code{timing}}{Numeric vector length one with the total elapsed time it took
  \code{\link{evolve_model}} to execute.}
}}

